+++ packages/vm/src/index.ts
@@ -1,13 +1,11 @@
 import { Account, Address, BigIntLike, toType, TypeOutput } from 'ethereumjs-util'
 import Blockchain from '@ethereumjs/blockchain'
 import Common, { Chain, Hardfork } from '@ethereumjs/common'
-import { StateManager, DefaultStateManager } from '@ethereumjs/statemanager'
-import { default as runCode, RunCodeOpts } from './runCode'
-import { default as runCall, RunCallOpts } from './runCall'
+import { StateManager, DefaultStateManager } from './state/index'
 import { default as runTx, RunTxOpts, RunTxResult } from './runTx'
 import { default as runBlock, RunBlockOpts, RunBlockResult } from './runBlock'
 import { default as buildBlock, BuildBlockOpts, BlockBuilder } from './buildBlock'
-import { EVMResult, ExecResult } from './evm/evm'
+import EVM from './evm/evm'
 import { OpcodeList, getOpcodesForHF, OpHandler } from './evm/opcodes'
 import { CustomPrecompile, getActivePrecompiles, PrecompileFunc } from './evm/precompiles'
 import runBlockchain from './runBlockchain'
@@ -74,7 +72,7 @@ export interface VMOpts {
    */
   common?: Common
   /**
-   * A {@link StateManager} instance to use as the state store (Beta API)
+   * A {@link StateManager} instance to use as the state store
    */
   stateManager?: StateManager
   /**
@@ -177,6 +175,11 @@ export default class VM extends AsyncEventEmitter {
   readonly blockchain: Blockchain
 
   readonly _common: Common
+
+  /**
+   * The EVM used for bytecode execution
+   */
+  readonly evm: EVM
 
   protected readonly _opts: VMOpts
   protected _isInitialized: boolean = false
@@ -311,7 +314,12 @@ export default class VM extends AsyncEventEmitter {
     }
     this.vmState = new VmState({ common: this._common, stateManager: this.stateManager })
 
-    this.blockchain = opts.blockchain ?? new (Blockchain as any)({ common: this._common })
+    this.evm = new EVM(this, {
+      common: this._common,
+      stateManager: this.stateManager,
+    })
+
+    this.blockchain = opts.blockchain ?? new Blockchain({ common: this._common })
 
     this._allowUnlimitedContractSize = opts.allowUnlimitedContractSize ?? false
 
@@ -419,28 +427,6 @@ export default class VM extends AsyncEventEmitter {
    */
   async runTx(opts: RunTxOpts): Promise<RunTxResult> {
     return runTx.bind(this)(opts)
-  }
-
-  /**
-   * runs a call (or create) operation.
-   *
-   * This method modifies the state.
-   *
-   * @param {RunCallOpts} opts
-   */
-  async runCall(opts: RunCallOpts): Promise<EVMResult> {
-    return runCall.bind(this)(opts)
-  }
-
-  /**
-   * Runs EVM code.
-   *
-   * This method modifies the state.
-   *
-   * @param {RunCodeOpts} opts
-   */
-  async runCode(opts: RunCodeOpts): Promise<ExecResult> {
-    return runCode.bind(this)(opts)
   }
 
   /**
