+++ packages/vm/tests/api/evm/precompiles/hardfork.spec.ts
@@ -21,7 +21,7 @@ tape('Precompiles: hardfork availability', (t) => {
     }
 
     let vm = await VM.create({ common: commonByzantium })
-    let result = await vm.runCall({
+    let result = await vm.evm.runCall({
       caller: Address.zero(),
       gasLimit: BigInt(0xffffffffff),
       to: ECPAIR_Address,
@@ -40,7 +40,7 @@ tape('Precompiles: hardfork availability', (t) => {
     }
 
     vm = await VM.create({ common: commonPetersburg })
-    result = await vm.runCall({
+    result = await vm.evm.runCall({
       caller: Address.zero(),
       gasLimit: BigInt(0xffffffffff),
       to: ECPAIR_Address,
@@ -61,7 +61,7 @@ tape('Precompiles: hardfork availability', (t) => {
 
     vm = await VM.create({ common: commonHomestead })
 
-    result = await vm.runCall({
+    result = await vm.evm.runCall({
       caller: Address.zero(),
       gasLimit: BigInt(0xffffffffff),
       to: ECPAIR_Address,
